# Conflicts:
#	src/main/java/godevenner/divexbackend/portfolio/controller/PortfolioController.java
#	src/main/java/godevenner/divexbackend/portfolio/repository/PortfolioEntryRepository.java
#	src/main/java/godevenner/divexbackend/portfolio/service/PortfolioEntryService.java
#	src/main/java/godevenner/divexbackend/portfolio/service/PortfolioEntryServiceImpl.java
#	src/test/java/godevenner/divexbackend/portfolio/controller/PortfolioControllerTest.java
#	src/test/java/godevenner/divexbackend/portfolio/service/PortfolioEntryServiceImplTest.java

+++ src/main/java/godevenner/divexbackend/portfolio/repository/PortfolioRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PortfolioRepository extends JpaRepository<Portfolio, Integer> {
+public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
 
     List<Portfolio> findByUserId(Long userId);
     Optional<Portfolio> findById(Long portfolioId);
