+++ src/main/java/godevenner/divexbackend/stock/repository/StockRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface StockRepository extends JpaRepository<Stock, Integer> {
+public interface StockRepository extends JpaRepository<Stock, Long> {
 
     Optional<Stock> findByTicker(String ticker);
